fix issues

+++ app/src/main/java/ir/co/tarhim/model/deceased/DeceasedDataModel.kt
@@ -1,10 +1,10 @@
 package ir.co.tarhim.model.deceased
 
 data class DeceasedDataModel(
-    val birthday: Int,
-    val deathday: Int,
+    val birthday: Long,
+    val deathday: Long,
     val id: Int,
     val imageurl: String,
     val name: String,
-    val timestamp: Int
+    val timestamp: Long
 )
