+++ app/src/main/java/ir/co/tarhim/model/deceased/DeceasedProfileDataModel.kt
@@ -1,8 +1,8 @@
 package ir.co.tarhim.model.deceased
 
 data class DeceasedProfileDataModel(
-    val birthday: Int,
-    val deathday: Int,
+    val birthday: Long,
+    val deathday: Long,
     val deathloc: String,
     val description: String,
     val imageurl: String,
